+++ frontend/app/src/theme/community/white.ts
@@ -1,4 +1,5 @@
-import { hexPercent, Theme } from "../themes";
+import type { Theme } from "../types";
+import { hexPercent } from "../themes";
 
 const textBox = "rgba(226,226,226,0.5)";
 const txt = "#242834";
