+++ frontend/app/src/theme/dark.ts
@@ -1,4 +1,5 @@
-import { hexPercent, Theme } from "./themes";
+import { hexPercent } from "./themes";
+import type { Theme } from "./types";
 
 const backgroundBlack = "#1B1C21";
